Allow for new AUX POW.
getauxblock is being depreciated and replaced by createauxblock and
submitauxblock. This allows for specifying a payment address when asking
for an auxblock instead of getting a wallet address so you don't have to
run a wallet where p2pool has access. Most coins support both methods,
some only support the old method (Terricoin is an example) and some only
support the new method (Myriadcoin is an example). Use `--merged` for
the old style and `--merged_addr` for the new style.